+++ Backend/models/OrderProduct.ts
@@ -39,7 +39,7 @@ OrderProduct.init(
       allowNull: false,
     },
     price: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
   },
